+++ NodeDependencyLookup/Editor/NodeHandler/AssetNodeHandler.cs
@@ -24,7 +24,6 @@ namespace Com.Innogames.Core.Frontend.NodeDependencyLookup
 		public int GetOwnFileSize(string id, string type, NodeDependencyLookupContext stateContext)
 		{
 			return 0;
-			//return NodeDependencyLookupUtility.GetPackedAssetSize(id);
 		}
 
 		public bool IsNodePackedToApp(string id, string type)