+++ changelog.md
@@ -1,6 +1,10 @@
+**1.2.3**
+ - Fixed issue with dependency resolvers getting updated several times when a dependency resolver gets activated/deactivated in the AssetRelationsViewer
+ - Sped up calculation for asset sizes by adding cache
+
 **1.2.2**
  - Added dependency resolver to resolve gameobjects dependencies inside the currently opened scene/prefab
- - Removed old ObjectDependencyResolver since its fully replaced by the ObjectSerializedDependencyResolver 
+ - Removed old ObjectDependencyResolver since its fully replaced by the ObjectSerializedDependencyResolver
 
 **1.2.0**
  - Added support for subassets